+++ checksum-QuickBEAM.Native.exs
@@ -1,8 +1,5 @@
 [
-  {"Elixir.QuickBEAM.Native-v0.10.18-x86_64-linux-gnu.so.tar.gz",
-   "sha256:80388bf4e69afc6aa7f3afabbbdc292e6506c4cdd0078c062c79d4f6984e784c"},
-  {"Elixir.QuickBEAM.Native-v0.10.18-aarch64-linux-gnu.so.tar.gz",
-   "sha256:4ff1087089932bd957a9bf547cf768668186095dc30cc5b590058b21c6cab29d"},
-  {"Elixir.QuickBEAM.Native-v0.10.18-aarch64-macos-none.so.tar.gz",
-   "sha256:5e20147c4720bbb0173771429bc489c81669c43464b9e79b06b636f466969cc6"}
+  {"Elixir.QuickBEAM.Native-v0.10.19-x86_64-linux-gnu.so.tar.gz", "sha256:PLACEHOLDER"},
+  {"Elixir.QuickBEAM.Native-v0.10.19-aarch64-linux-gnu.so.tar.gz", "sha256:PLACEHOLDER"},
+  {"Elixir.QuickBEAM.Native-v0.10.19-aarch64-macos-none.so.tar.gz", "sha256:PLACEHOLDER"}
 ]